+++ plugins/ZmqSubscriber.cpp
@@ -38,6 +38,7 @@ public:
       for (auto& conn_string : m_connection_strings) {
         try {
           m_socket.disconnect(conn_string);
+          m_connection_strings.erase(conn_string);
         } catch (zmq::error_t const& err) {
           ers::error(ZmqOperationError(ERS_HERE, "disconnect", "receive", err.what(), conn_string));
         }
@@ -48,15 +49,18 @@ public:
 
   std::string connect_for_receives(const nlohmann::json& connection_info) override
   {
+    std::set<std::string> new_connection_strings;
     if (connection_info.contains("connection_string")) {
-      m_connection_strings.insert(connection_info.value<std::string>("connection_string", ""));
+      if (m_connection_strings.count(connection_info.value<std::string>("connection_string", "")) == 0)
+        new_connection_strings.insert(connection_info.value<std::string>("connection_string", ""));
     }
 
     for (auto& conn_string : connection_info.value<std::vector<std::string>>("connection_strings", {})) {
-      m_connection_strings.insert(conn_string);
+      if (m_connection_strings.count(conn_string) == 0)
+        new_connection_strings.insert(conn_string);
     }
 
-    if (m_connection_strings.size() == 0) {
+    if (m_connection_strings.size() == 0 && new_connection_strings.size() == 0) {
       throw ZmqOperationError(ERS_HERE, "resolve connections", "receive", "No valid connection strings passed", "");
     }
 
@@ -68,10 +72,11 @@ public:
         throw ZmqOperationError(ERS_HERE, "set timeout", "receive", err.what(), *m_connection_strings.begin());
       }
     }
-    for (auto& conn_string : m_connection_strings) {
+    for (auto& conn_string : new_connection_strings) {
       try {
-          TLOG_DEBUG(19) << "Connecting to publisher at " << conn_string;
+        TLOG_DEBUG(19) << "Connecting to publisher at " << conn_string;
         m_socket.connect(conn_string);
+        m_connection_strings.insert(conn_string);
       } catch (zmq::error_t const& err) {
         ers::error(ZmqOperationError(ERS_HERE, "connect", "receive", err.what(), conn_string));
       }
